solo era poner en xlsx

+++ database/migrations/2024_10_23_204546_create_asientos_table.php
@@ -36,7 +36,7 @@ return new class extends Migration {
             $table->string('consecutivo');
             $table->string('periodo');
             $table->string('plan_cuentas');
-            $table->integer('numerounico')->nullable();
+            $table->bigInteger('numerounico')->nullable();
             $table->timestamps();
         });
     }
